Interview Section, Job seeker Profile View, Admin Validation

+++ app/Http/Controllers/Admin/OppertunitiesController.php
@@ -3,12 +3,13 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Admin\AdminBaseController;
 use Illuminate\Http\Request;
 use App\Models\Admin\Oppertunities;
 use App\Models\Admin\Companies;
 use Illuminate\Support\Facades\Auth;
 
-class OppertunitiesController extends Controller
+class OppertunitiesController extends AdminBaseController
 {
     /**
      * Create a new controller instance.
@@ -28,6 +29,9 @@ class OppertunitiesController extends Controller
     
     public function index()
     {   
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
         $fetch = Oppertunities::join('companies', 'companies.id', '=', 'oppertunities.company_id')
                 ->get(['oppertunities.*', 'companies.name as company_name']);
         $data['oppertunities'] = $fetch;
@@ -36,12 +40,18 @@ class OppertunitiesController extends Controller
 
     public function add()
     {
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
          $data['companies'] = Companies::where('status','1')->get();
         return view('admin/oppertunities/add',$data);
     }
 
     public function edit($id)
     {
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
         $data['companies'] = Companies::where('status','1')->get();
         $data['oppertunity'] = Oppertunities::where('id', $id)->first();
         return view('admin/oppertunities/edit',$data);
@@ -49,6 +59,9 @@ class OppertunitiesController extends Controller
 
     public function view($id)
     {
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
         $fetch = Oppertunities::join('companies', 'companies.id', '=', 'oppertunities.company_id')
                 ->get(['oppertunities.*', 'companies.name as company_name'])->where('id',$id)->first();
         $data['oppertunity'] = $fetch;
@@ -57,7 +70,9 @@ class OppertunitiesController extends Controller
 
     public function store_oppertunity(Request $request)
     {
-
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
         $validated = $request->validate([
             'title' => 'required|max:255',
         ]);
@@ -86,7 +101,9 @@ class OppertunitiesController extends Controller
 
     public function update_oppertunity($id, Request $request)
     {
-
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
         $validated = $request->validate([
             'title' => 'required|max:255',
         ]);
@@ -113,6 +130,9 @@ class OppertunitiesController extends Controller
     }
 
     public function delete_oppertunity($id){
+        if(!$this->check_role()){
+            return redirect()->route('home');
+        };
         Oppertunities::where('id',$id)->delete(); 
         return redirect()->route('admin.oppertunities')
         ->with('success','Oppertunity Deleted successfully.');
